requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
chromadb==0.4.22
pypdf2==3.0.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
sentence-transformers==2.2.2

